+++ packages/capabilities/src/index.js
@@ -3,11 +3,12 @@ import * as Top from './top.js'
 import * as Store from './store.js'
 import * as Upload from './upload.js'
 import * as Voucher from './voucher.js'
+import * as Access from './access.js'
 import * as Utils from './utils.js'
 
-export { Space, Top, Store, Upload, Voucher, Utils }
+export { Access, Space, Top, Store, Upload, Voucher, Utils }
 
-/** @type {import('./types').AbilitiesArray} */
+/** @type {import('./types.js').AbilitiesArray} */
 export const abilitiesAsStrings = [
   Top.top.can,
   Space.space.can,
@@ -24,4 +25,7 @@ export const abilitiesAsStrings = [
   Store.list.can,
   Voucher.claim.can,
   Voucher.redeem.can,
+  Access.access.can,
+  Access.authorize.can,
+  Access.session.can,
 ]