+++ packages/upload-client/test/helpers/block.js
@@ -5,6 +5,6 @@ import { sha256 } from 'multiformats/hashes/sha2'
 /** @param {Uint8Array} bytes */
 export async function toBlock(bytes) {
   const hash = await sha256.digest(bytes)
-  const cid = CID.create(1, raw.code, hash)
+  const cid = CID.createV1(raw.code, hash)
   return { cid, bytes }
 }